feat(SvcApiGateway): Refactor aws clients and cdk builders
