+++ lib/libs/cdk-builders/DynamoBuilder.ts
@@ -1,9 +1,9 @@
 import { Construct } from "constructs";
 import * as dynamodb from "aws-cdk-lib/aws-dynamodb";
-import { createResourceNameId } from "../helpers";
+import { createResourceNameId } from "../../stacks/svc-api-gateway-stack/cdk/helpers";
 import { RemovalPolicy } from "aws-cdk-lib";
 import { NodejsFunction } from "aws-cdk-lib/aws-lambda-nodejs";
-import { DynamoTableNames } from "./types";
+import { DynamoTableNames } from "../../stacks/svc-api-gateway-stack/cdk/dynamo/types";
 
 export type createDynamoOptions = {
   tableName: DynamoTableNames;
