+++ lib/libs/cdk-builders/LambdaBuilder.ts
@@ -1,14 +1,14 @@
 import { NodejsFunction } from "aws-cdk-lib/aws-lambda-nodejs";
 import { Construct } from "constructs";
-import { createResourceNameId } from "../helpers";
-import { GlobalEnvironmentVars } from "../../../../libs/environment";
+import { createResourceNameId } from "../../stacks/svc-api-gateway-stack/cdk/helpers";
+import { GlobalEnvironmentVars } from "../environment";
 import path = require("path");
 import { Runtime } from "aws-cdk-lib/aws-lambda";
-import { LambdasFunctionNames } from "./types";
+import { LambdasFunctionNames } from "../../stacks/svc-api-gateway-stack/cdk/lambda/types";
 
 type createLambdaFunctionOptions = {
   lambdaName: LambdasFunctionNames;
-  fileNameImlCode: string;
+  pathStackHandlerCode: string;
   environment: Record<string, string>;
 };
 
@@ -27,17 +27,14 @@ export class LambdaBuilder {
    * @returns {NodejsFunction}
    */
   public createNodeFunctionLambda(options: createLambdaFunctionOptions) {
-    const { lambdaName, fileNameImlCode, environment } = options;
+    const { lambdaName, pathStackHandlerCode, environment } = options;
     this.lambdaFunctions[lambdaName] = new NodejsFunction(
       this.scope,
       createResourceNameId(lambdaName),
       {
         runtime: Runtime.NODEJS_18_X,
         handler: "handler",
-        entry: path.join(
-          __dirname,
-          `../../services/lambda-api-integrations/management/${fileNameImlCode}`
-        ),
+        entry: path.join(__dirname, `../../stacks/${pathStackHandlerCode}`),
         environment: {
           DEPLOY_ENVIRONMENT: GlobalEnvironmentVars.DEPLOY_ENVIRONMENT,
           LOGGER_LEVEL: GlobalEnvironmentVars.LOGGER_LEVEL,
