+++ lib/libs/clients/dynamodb/models/management.ts
