+++ lib/libs/clients/dynamodb/operations/helpers.ts
