+++ lib/libs/clients/dynamodb/operations/types.ts
