+++ lib/libs/clients/dynamodb/services/index.ts
