+++ lib/libs/clients/dynamodb/tables/table-search-columns.ts
