+++ lib/libs/clients/dynamodb/tables/tables-index.ts
