+++ lib/stacks/svc-api-gateway-stack/cdk/api/management.ts
@@ -1,4 +1,4 @@
-import { ApiRestBuilder } from "./ApiRestBuilder";
+import { ApiRestBuilder } from "../../../../libs/cdk-builders/ApiRestBuilder";
 import { buildCorsConfigurations } from "./cors";
 import { ResourcesAPI } from "./resources/types";
