+++ lib/stacks/svc-api-gateway-stack/cdk/dynamo/management.ts
@@ -1,12 +1,12 @@
 import { AttributeType } from "aws-cdk-lib/aws-dynamodb";
-import { DynamoBuilder } from "./DynamoBuilder";
+import { DynamoBuilder } from "../../../../libs/cdk-builders/DynamoBuilder";
 import { DynamoTableNames } from "./types";
 
 export const buildDynamoConstructs = (builder:DynamoBuilder) => {
   builder.createTable({
     tableName:DynamoTableNames.Users,
     partitionKey: {
-      name: "user_id",
+      name: "userId",
       type: AttributeType.STRING,
     },
   });
