+++ lib/stacks/svc-api-gateway-stack/index.ts
@@ -1,43 +1,11 @@
-/* import { Construct } from "constructs";
-import * as cdk from "aws-cdk-lib";
-import { GeneralApiGateway } from "./cdk/api/general-api-gateway";
-import { ManagementLambdas } from "./cdk/lambda/management-lambdas";
-import { manageMentResources } from "./cdk/api/resources/management/resources";
-import { ManagementDynamoDB } from "./cdk/dynamo/management-dynamo";
-import { ManagementDynamoKeyName } from "./cdk/dynamo/types";
-import { ManagementLambdaKeyNames } from "./cdk/lambda/types";
-
-export class SvcApiGatewayStack extends cdk.Stack {
-  constructor(scope: Construct, id: string, props?: cdk.StackProps) {
-    super(scope, id, props);
-
-    // constructs
-    const managementLambdas = new ManagementLambdas(this);
-    const apiGateway = new GeneralApiGateway(this);
-    const managementDynamoDB = new ManagementDynamoDB(this);
-
-    // Api gateway settings
-    apiGateway.setLambdaHandlers(managementLambdas.getLambdaHandlers);
-    apiGateway.addApiResourceFromRoot({ resources: manageMentResources });
-
-    // DynamoDB settings
-    managementDynamoDB.setLambdaHandlers(managementLambdas.getLambdaHandlers);
-    managementDynamoDB.grantWritePermissionsToLambdas({
-      keyNameTable: ManagementDynamoKeyName.Users,
-      keyNameLambdas: [ManagementLambdaKeyNames.GetUsers],
-    });
-  }
-}
- */
-
 import { App } from "aws-cdk-lib";
-import { GenericStack } from "./cdk/GenericStack";
-import { ApiRestBuilder } from "./cdk/api/ApiRestBuilder";
+import { GenericStack } from "../../libs/cdk-builders/GenericStack";
+import { ApiRestBuilder } from "../../libs/cdk-builders/ApiRestBuilder";
 import { buildApiRestConstructs } from "./cdk/api/management";
 import { createManagementApiResources } from "./cdk/api/resources/management/resources";
-import { DynamoBuilder } from "./cdk/dynamo/DynamoBuilder";
+import { DynamoBuilder } from "../../libs/cdk-builders/DynamoBuilder";
 import { buildDynamoConstructs } from "./cdk/dynamo/management";
-import { LambdaBuilder } from "./cdk/lambda/LambdaBuilder";
+import { LambdaBuilder } from "../../libs/cdk-builders/LambdaBuilder";
 import { buildLambdaConstructs } from "./cdk/lambda/management";
 import { LambdasFunctionNames } from "./cdk/lambda/types";
 import { DynamoTableNames } from "./cdk/dynamo/types";
@@ -62,9 +30,7 @@ export const buildSvcApiGatewayStack = (app: App) => {
   // Dynamo permissions settings
   dynamoBuilder.grantWritePermissionsToLambdas({
     dynamoTable: DynamoTableNames.Users,
-    lambdas: [
-      lambdaBuilder.getLambdaFunctions[LambdasFunctionNames.GetUsers],
-    ],
+    lambdas: [lambdaBuilder.getLambdaFunctions[LambdasFunctionNames.GetUsers]],
   });
 
   // ApiRest settings
