--- lib/stacks/svc-api-gateway-stack/services/clients/dynamodb/tables/tables.ts
@@ -1,9 +1,0 @@
-export enum ManagementTables {
-    'Users'='Users'
-}
-
-export enum TrenLigeroTables{
-    'Trenes'='Trenes'
-}
-
-export type DynamoTables = ManagementTables | TrenLigeroTables
