+++ lib/stacks/svc-api-gateway-stack/services/handlers/management/get-dependencies.ts
