+++ lib/stacks/svc-api-gateway-stack/services/handlers/types.ts
@@ -1,4 +1,4 @@
 export interface QueryParamsPagination {
   page: string;
-  page_size: string;
+  pageSize: string;
 }
